Allow user changes with autogen email to be saved
Previously autogenerated emails in the format NNNNNN-provider would not
allow a user to be saved since a requirement for a user is a valid
email address.
This meant that the token for validation emails would not be saved to
the database making it impossible to verify the new email.

Validation now has been changed to recognise anything ending in
-facebook, -twitter or -google as valid emails for the purpose of
saving the user so that the token can be saved.