Fix a bug with cancelling reservations
When a reservation is requested, the target user is not added to the event's group. If the user then cancels the reservation, the API attempts to change their membership status. However, no such membership exists.

This fix does not change the membership status if the user is cancelling a reservation.