Skip over deleted users for events booking CSV.
Deleted users cannot be retrieved by the UserAccountManager if they have been deleted, yet we were doing just that and re-throwing the exception while building the CSV. This commit catches the exception and advances the loop to skip over the student that was deleted.

We could possibly want to add a line that says DELETED or something like that.