Add a server property to allow overriding of the question misuse threshold
This property is only read at startup, and so can't be changed on the fly.
It wouldn't be too difficult to change to reading it on the fly, but it
felt more efficient to parse the string only once, not every single
time a question is attempted on the platform.