Retrieve group progress assignments sorted by due date and set date.
This sorts assignments by due date (first) and by creation date (second).

This may be the stupid way of doing this but it's also the most legible. Sorting on ordered streams in Java is stable, but writing the predicate might have been more work than it was worth. Divide and conquer wins the day.