Check that users creating groups or setting assignments are teachers
Historically we never intended these features to be teacher-only, and in
the many years since we changed our minds, nobody has actually changed the
code to enforce the rules. Until now with a small amount of JavaScript, it
was possible to trick the frontend into showing you teacher functionality
and the API would happily let you.
We should review whether you should be allowed to access existing groups
should your account be demoted, but this prevents new assignments and new
groups for student users.