Log the referer header when image files are not found
This can help work out what page a file is referenced on much more easily.
I also @Deprecated the IsaacController method that does all the work, since
it takes a version parameter but does not use it. We do not directly access
this endpoint, but call it internally from the main image endpoint in the
SegueContentFacade.