Refactor MailJet API into wrapper class and improve exception handling
This moves much of the complexity out of the ExternalAccountManager, but is
still not particularly generic. I think the MailJet flow may be too MailJet
specific to warrant the additional exception classes and interfaces needed
to fully generify it. This seems an acceptable compromise.
Rename some methods, add JavaDoc, and move the SubscriptionAction enum
into its own class.