Enable the numeric validator to correctly parse "10^4" as a number
The previous regex only worked if the power of ten was prefixed by a number
multiplying it, and cannot be modified to allow bare powers of ten since
for bare powers of ten we need to replace with "1e..." and not "e...".
Add some tests for the new example case.