Prevent over-caching of quizzes by the GitContentManager
After the move to using "latest"/"live" for the index names, we have to use
the weird `getCurrentContentSHA()` call to pass into the content manager.
If you check all the other `contentManager.getContentDOById()` calls then
you will see that this is the correct way to call it.
Why is that parameter necessary at all? Legacy code, I guess!