Add most recent questions and oldest incomplete questions to progress stats
To keep the response size manageable, only include the most recent 5
questions and oldest 5 incomplete questions. Use a FIFO queue to keep the
most recent 5 questions, since this is more efficient than a list.
Doing this assumes the input Map is sorted, which it was not despite the
results being returned from Postgres in order. LinkedHashMaps offer this
order without much of an overhead and required minimal code change!

I also implemented my 4 year old TODO and removed the need for a local
GameManager by using the static method and the existing DTOs, which should
also be a performance boost by not loading everything from ElasticSearch
twice.