Strip square brackets from IPv6 addresses in RequestIPExtractor
Future versions of Jetty return IPv6 addresses in URL form, wrapped in
square brackets like "[::1]" instead of "::1", which Postgres does not
accept. Since this currently only affects local APIs, and is safe, I add
this in now rather than when we upgrade Jetty.
The code snippet to do this stripping was based on how the Jetty source
used to do this stripping itself.