Simplify cookie JSON property names
Since the new cookie format is not only incompatible with the old format,
but the updated library cannot even parse the old cookies, take the most
of the opportunity to shorten and simplify the cookie JSON key names.
We will have to force-logout everyone to migrate the cookie formats anyway,
so this change is no additional overhead.