Allow multiple boolean search clauses to be matched against the same field
To get the question filter to match questions with "statics" or "dynamics" and  of type "problem_solving" or "books" we need two separate clauses on the "tags" field.
To achieve this I created a BooleanSearchClause record class (rather than including a tuples library) and told ElasticSearchProvider how to construct a bool match query from a list of these records.
That method is a lot more straightforward now and even the nested queries can be simplified by only treating them differently just before adding the clause to the master query.
I then followed the compiler's complaints to update the fieldsToMatch datastructure to the new list of BooleanSearchClauses.

It would have been slightly nicer to review if I would have remembered to commit between the refactoring and the fixing of generateBoolMatchQuery, sorry!

I have tested our major routes to search content quite thoroughly and as the logic is now easier to comprehend I'm confident it is correct.
During the transforming of the clause data structure, the only section of work that was not checked by the compiler was making sure the correct boolean operatorwas copied over - I have manually double checked those and they are fine.