Upgrade Orika to version 1.5.4
This is an attempt to fix the issues with the mappers not initialising
themselves correctly. The conversion methods now take an extra argument
which is not well documented; it _seems_ to be safe to ignore it as I have
done here, but I am not sure how safe this is.