Use the default feedback for question if no better feedback is generated
If it exists, and no better feedback has been generated in the validator,
use the default feedback from the question. This is very straightforward
for all question types except numeric which will be another commit.
Refactor the ChoiceQuestionValidator (used for multi-choice) to fit the
newer style of having a feedback object, so it can have the same solution
as the rest too.