Add subset validation of ItemChoice questions
This required adding `allowSubsetMatch` to `ItemChoice`, a flag determining whether the user's submitted answers can be a subset of one of the question choices

If a user ticks options {1, 3} and the answer is {1, 2, 3} and `allowSubsetMatch` is `true`, then the user's submission should be marked as correct. I overrode `getOrderedChoices`
so that the ordering of the choices makes sense if some are subset match and some aren't - the 'strict' (non-subset match) choices are ordered before the subset match ones.