Restore old desired behaviour to Orika Custom and Bidirectional mappers
After updating to v1.5.x, the default behaviour of mappers became much
stricter and did not correctly map all of our polymorphic subclasses in
all cases.
This issue is the one described in orika-mapper/orika#171, and the fix
is to override canConvert for CustomConverter and BidirectionalConverter
with the old code as suggested in that issue.

I can't be sure that _all_ of these Converters need to be the polymorphic
supporting version - but Item, Choice and QuestionValidationResponse
definitely need to be for questions to continue functioning.