Add a displayUnit to numeric questions which overrides requiresUnits
The aim is that the displayUnit gets shown to the user and so we do not
need to validate the units since they are not asked for. Guarding against
cases where units have been specified anyway makes up most of the changes.
Two new content errors are added to highlight these issues.

I also removed sig figs from the DTO, since they were JsonIgnored and they
were not being used and it simplified the file a lot.