Ensure that the SMTP settings configured per-email are actually used
Previously we used Session.getDefaultInstance which only initialised using
the provided properties once and after that ignored the provided properties
entirely. Creating a new Session object for _every_ email is very expensive
however, so we can cache it because the properties don't differ that much
between our emails.
This is a prerequisite for the overrideEnvelopeFrom email property to be in
any way effective, which we are going to use to route emails in Exim.