Make the GuiceConfigurationModules factories for their own injectors
This is one, albeit not that nice, way of ensuring only one Guice injector
is created. It avoids an extra factory class by making the configuration
module the factory for the injector.
The ETL injector was only created twice, but I still made it use the same
pattern for consistency.

In theory, this makes all the static singletons in the configuration
modules redundant because there is only one instance of the injector and so
just being a singleton (without the need for the static pattern) is enough.
But if we didn't want overengineering, we wouldn't be using Java ¯\_(ツ)_/¯