Prevent Quartz service from starting if database is read-only
Quartz relies on being able to write to the database to function, and so
cannot run if the database it is connected to is a streaming replica.
This adds a new function to the database class to allow checking of the
status before attempting Quartz initialisation.

This will not start Quartz if the database becomes the writable primary
node at a later time.