Add hiddenFromRoles property to quizzes to allow more flexible filtering
Currently there is only "visibleToStudents", this is the first step to
deprecating that property. The changes attempt to be backwards compatible;
if the content does not have the new property but only the old one (as it
must for one release cycle) then the old behaviour should be unchanged.
We will then need to remove the old property from the content and then from
the DOs and DTOs.