Tear-up and redesign `updateGameboard` endpoint
I've renamed and refactored this endpoint so its name is indicative of its useage. The original endpoint could only do what this endpoint does, so this change loses no functionality at all (especially considering that endpoint hasn't been used in the front-end, presumeably since the React redesign).
Most of the code is the same, with all the same checks, minus those to require the new DTO to be equal to the old one, since this endpoint only takes a new title now.