Mark correct numeric units as correct even for a known incorrect choice
This will only work in some circumstances:
 - the units provided match a known correct answer.
 - the value and units both match a known incorrect answer.
Then, and only then, can we safely mark the units as correct and the value
as incorrect. Limited in usefulness, but addresses a noticed issue in the
content without (I hope!) any side effects.
Also add a test for the noticed case for future reference.