Wrap PreparedStatements and ResultSets in try-with-resources
To avoid resource leaks, these classes should be closed after use, which
the try-with-resources block will automatically do correctly.
Apparently calling pst.setXXX(YYY) inside the try section is dangerous
because the exception might not be caught correctly? I avoided it, but
doing so means that ResultSets often need their own nested try blocks and
the code is slightly less readable than it was (and it was bad!).

Some methods were too complex to do this for without major refactoring and
these have "//FIXME: try-with-resources!" lines added.