Modify validate methods to conditionally ignore significant figures
The validateWith/WithoutUnits methods determine the number of
significant figures the user answer and true answer should be
compared to, and then round both to this length before comparison.
This change means these two steps do not happen when
disregardSignificantFigures for the question is set, and the
answers are compared at their original precision.