Stop augmenting bulk gameboard requests with saved-to-user information
When you load a single gameboard, it makes sense to ask whether it is saved
to the current user's account. When you load multiple gameboards, we do not
use this information at present and adding it requires a separate database
query per board. Removing this augmentation in this case is a quick win.