Alter updateStatus to use transactions and fix tests
There are several cases for updating a booking, and the previous code
was a mess of LGTM warnings and unclosed PreparedStatements. There are
four cases, so i just wrote four separate private methods to be clear
about what is happening and to avoid concatenating SQL statements.

The tests were tricky to fix, because some of them used mocks they
didn't declare in replay or verify. Now there is an array of mocked
objects to always replay in every test to be safe!