Add SCrypt as default password hashing algorithm
Use the BouncyCastle library to provide the SCrypt algorithm, and choose
strong (perhaps too strong?) parameters for it. I implemented it as a
versioned subclass just in case we want to adjust the parameters after
deployment and thus need a nice naming convention.

On isaac-1, the algorithm manages 2.4 H/s with N=65536 (i.e. 420ms),
but would manage 4.8 H/s with N=32768 (i.e. 209ms) which is the
recommended value from 2017. We may want to adjust to the faster one
once I have checked the cracking speeds of these, but the slow choice
in this commit is the same speed as the existing PBKDF2 algorithm.