Make the PgEventBooking a POJO and remove update logic
The object should just be a store of data, not something that holds a
database connection and can update itself. This removes a
(non-functional! the column names are all wrong!) method for the object
to update itself, and adds instead a new manager method to get a booking
by ID.