Use separate web.xml files for local and production
Remove localhost from the CORS domains for production, which will
prevent the development version of the content editor from accessing
staging but tightens up the allowed origins.
Allow any 4-digit localhost port to access the local version of the API.
The local configuration is a copy of the old API one but with more ports
allowed on localhost.