Stop using hashing algorithm to generate locked account credentials
The "computeHash" method shouldn't be public, for hashing algorithms,
and in order to make it private this code needed rewriting.
The "LOCKED" rows added to user_credentials do not need to be valid
hashes; so long as the output is not a valid hash the security comes
from the fact no hash can match that value anyhow. So instead of the
expensive hash being used, just add a random but impossible-to-match
placeholder.
The reset token does want to be more secure, and so I left that using
the hashing algorithm (and hashPassword does the same thing as
computeHash but with the Base64 and String added in!).