Make the computeHash method of hashing algorithms private
For the core password hashing and storage this was always only internal
anyway. In the previous commit I removed its use in generating reset
tokens since that was unnecessary.
This way we consistently ensure passwords are stored in the same Base64
format.
This enables new algorithms to be added that do not have an obvious
"computeHash" output; chained/compound hashing algorithms.