Use ORs not regex when querying by multiple question IDs
This produces horrific and long queries, but it is at least 10x faster
for all the cases I have tested and as much as 30x faster for big
queries.
The repeated OR clauses allow the database to perform much better index
scans and avoid the need to filter results matching every question part
ID against a (very large) regex.
I have tested it with 800 question IDs, and there it is much faster than
the 10k character long(!) regular expression generated before.