Allow GameboardManager to augment using only relevant questions
Previously the My Assignments endpoint had to load all the question
attempts first. Other places got round this by taking the contents of
the gameboards and getting just the relevant questions; but we couldn't
do that here because we don't have the gameboards outside getGameboards
in the first place. Thus we have to do the augmenting inside a new but
very similar method.
This isn't the best way to do this (just adding a new method), but the
whole lot needs refactoring anyhow.