Skip question ID filtering when loading by very many IDs
The database queries become very inefficient somewhere between 150-300
OR clauses; it loads the rows by user and then compares each row to the
filter clause. Instead in these cases just load all the attempts into
Java; we add them to a HashMap and querying from this is efficient.
This is a tradeoff, and MAX_PAGE_IDS_TO_MATCH may need adjusting.