Don't enable local profile by default and don't filter resources
This fixes a bug in d035e0017615c915ca5b625cf65d9a8df0f63334 whereby the
production build, even though it didn't specify the local profile, was
using the local profile.
If we don't use the local profile, we don't use the resource filtering
that it has configured. ETL works fine without this, and it seems to
have been added in 1f491a33e22f7bcbdcb408f89ed809513572535f back when
we did use filters. We do not in the Jetty world.
I have tested that nothing else is changed inside the WAR file when I do
this beyond the expected web.xml changes, so the filters were not
changing anything as ETL's build suggested.