Add exclusion filter for deprecated content to board generation query
This commit adds the deprecated filter to the query used by the /gameboards
endopoint to select random questions, which was returning deprecated content to
the Question Finder. In other places this filtering is done on the frontend,
however it isn't straightforward to do this the same way for the Question
Finder. This solution works but should probably be re-visited for consistency.