Use Maven Failsafe to run integration tests
This commit makes a few changes to how tests are run. The need for this became
apparent when the CI build agent was running out of memory when running the new
integration tests (which spin up docker containers). This was related to
PowerMock causing IsaacIntegrationTest's static initialisation code to run
twice, doubling the number of docker containers created unecessarily.

It seems important to be able to use PowerMock in these tests if the need
arises, so the static initialisation was replaced with a @BeforeClass method.
This has the added benefit of eliminating leftover state between test classes as
the containers are re-created for every test class.

Beside being slower, this creates a potential memory issue unless these tests are
run sequentially (the existing unit tests are not). In order to configure the
unit tests and integration tests separately Failsafe was introduced.

The "fork" (i.e., how many JVMs are created to run tests in parallel) options of
the existing unit tests were adjusted to re-use JVMs rather than creating new
ones, which speeds the suite up by over 50%. With these changes, the time taken
to run both suites is around 2 minutes, similar to the prior state.

It's worth noting that Failsafe runs the integration tests as part of the verify
Maven target, not the test target. The test target will only run the unit tests.