Add extra request class and fix names in facade tests
We're going to need both a Request and an HttpServletRequest for one
of the methods soon, so add in the plain Request and rename the existing
one for clarity.
Since Requests are used only for caching, which we don't care about in
these tests, use a NiceMock and not a plain Mock to ignore method calls.