Fix overly aggressive caching behaviour for available tests
Prior to this change, tests were hard cached for an hour, always.
I swapped it to NEVER_CACHE_WITHOUT_ETAG_CHECK to enable checking
the ETag, but since we never generated a 304 this was pointless.
Adding in the ETag check (via generateCachedResponse) fixed this, but
still left the issue that if the startIndex was specified, or if the
user's role changed (i.e. on a shared browser during testing),
then we used the incorrect cached value anyway.
So finally I updated the ETag to include the role and startIndex too.
Adding the values together is safe, since ETag collisions are unlikely
and even malicious adjusting of startIndex still can't leak any new
data, only generate a useless "304 Not Modified".