Rewrite cloze validator to support blanks and subset matching
Previously the validator didn't correctly support subset matching
(in fact it was badly broken, using sets which did not preserve the
order of the items submitted!), and did not allow question setters to
subset match with blanks.
Now an answer can be specified as (id1, NULL_ID, NULL_ID) and marked
as allowing subset matches, and then submissions such as
(id1, id2, id3) or (id1, id3, NULL) could match it. This should be
useful for adding common wrong answer feedback.

There is an open question about whether subset matching should allow
empty submissions from users to match (provided the entire answer is
not empty), but at present this is not allowed.