Add an index on user group_membership status
We query users' membership status of groups frequently, and often
PgUserGroupPersistenceManager::getGroupMembershipList would perform a
sequential scan in the production database. This is wildly inefficient,
and the filtering was removing 100k+ rows. Adding the index adds a 50%
size overhead (the table is 25MB, the new index an extra 13MB) but
makes the query run in 0.1ms instead of 38ms and reduces disk reads
dramatically.

As always, I recommend using something like:
   docker exec phy-pg-live pg_dump -n public -s -U rutherford rutherford
then removing functions and unnecessary changes, to update the schema.