Allow detailed item feedback in cloze questions
It is opt-in, via the detailedItemFeedback property of Cloze Questions.

The only nasty thing is that the QuestionValidationResponseDeserializer
doesn't have any type information; we don't store that in the question
attempts table. That means it has to work out what to deserialise
something into based on the JSON content. For QuantityVRs, we use the
fact that the submitted answer will be a Quantity; but for cloze
questions we re-use ItemChoices and so we cannot do that. Instead just
check if the unique property of ClozeVRs is set.