Use a SegueErrorResponse for unhandled exceptions
In RESTEasy 6.1.0, they implemented the changes in Jakarta that require
a default exception handler, but their new default handler didn't seem
to work in the way the old one did. But I cannot work out what the old
one actually did.
This ensures that unhandled exceptions that escape our code will be
returned to the client inside a SegueErrorResponse.