Never return null retrieving gameboards by IDs
Perhaps there is something to be said for "null in, null out", but if
you pass the empty list in, you should get the empty list out and not
a null value.
I'd rather consistency (never return null, just an empty list on
invalid input) though.

Remove the now-unnecessary null-avoiding code elsewhere too.