Fix mishandling of CORS OPTIONS requests
Jetty has a CORS filter that automatically responds to OPTIONS requests,
but it then passes the response and request down to RESTEasy to allow
the application to override it on a per-endpoint basis if desired.
We have nothing tagged @OPTIONS, and so these requests cause a
DefaultOptionsMethodException to be thrown which then gets ignored and
turned into the valid response by the default error handler.

However, the new error handler added in 8cbd09d now catches this
exception and turns it into a 500 rather than a CORS response.
This breaks CORS. So far as I can tell, just disabling the passing of
OPTIONS requests down to the application fixes this issue and so that
is all this commit does. We had not used the ability to override these
OPTIONS responses anywhere.

There remains a risk that there are other exceptions used to manage
behaviour in RESTEasy, which leaving the handler as-is might break,
but I have tested things more thoroughly now and am happy we won't
hit those cases.