Add a new validation response type for Item feedback
As noted in the Javadoc, this new validation response type is named
after Item questions but is unlikely to be useful for them directly.
Instead it is needed for Cloze questions, which inherit from Item
questions, and named generically in case it may later prove useful for
Parsons questions also.