Ensure "latest" alias gets stored on-disk by ETL
Previously there were two very similar code paths for the "manual alias"
update, and the queued "latest version" update. But the latest version
path didn't store the alias index on-disk, and so it was a long way
behind the actual latest version and causing ETL to index old content
each startup.
This should be safe, since writing the property file is a synchronised
method and shouldn't clash between the two threads.