Fix integer overflow issue with MailJet user IDs
The JSON library is clever enough to use Integers for short numbers and
Longs for larger ones, but when I wrote the code we clearly only had
small enough numbers so I didn't realise this. Fortunately Integers are
all a subset of Long, so we can just use Long here.