Add an endpoint to shutdown the Quartz scheduler
The way we use Quartz means that we cannot restart it again after
shutting it down this way, so the endpoint makes it clear that the
operation is permanent.
Sadly the "isStarted" method just checks that the scheduler has ever
been started, not whether is has subsequently been shutdown. I altered
the existing methods to better reflect this.