Escape underscores in question IDs when using LIKE
When we use LIKE matching to find question IDs based on the page ID, we
do `question_id LIKE 'page_id%'`. However, the underscore character
commonly found in our page IDs is treated as a wildcard, which means
that the index scan looks for `>= page` and `< pagf` rather than using
the full intended `>= page_id` and `< page_ie`!

Coupled to the fact that each set of book questions shares a common
prefix (e.g. `gcse_`), this means that the matches were needlessly
broad and the subsequent filtering was then having to remove many rows
and pages that could easily have been excluded in the index scan.

Adding this escaping leads to a ~50% speedup for a common query for our
most active users.