Set SEGUE_AUTH_COOKIE as SameSite=Strict
I think this is fine; the difference between Lax (what we have been
implicitly using before now) and Strict is about top-level navigations
and third-party triggered GET requests.
We don't use any cookies on the very-top-level GET; that goes to the
frontend which is stateless.
I also do not think that third-party origins should be making
authenticated GET requests (our CORS would stop the response being used
in most cases anyhow), so that should also be safe.

If this breaks, moving to SAME_SITE_LAX_COMMENT is the next one to try.