Swap ES search of 'value' and 'children' for 'searchable_content' field
Searching on "value" and "children" only tries to match the search query against the value and children fields on the top level of the content document.
On pages, these are empty and or a list of content objects so never aid in matching anything useful.
I'm quite sure inital implementation assumed it would match any value in the children recursively.
To correct for this assumption we created a searchable_content field which is populated during ETL by walking the content collecting titles and values.
I'm not sure how I missed this search endpoint while fixing it for the other(s) a year or so ago.