Improve bulk actions for group management
This adds two new database methods for loading bulk group information:
one to bulk load groups themselves, and one to load a user's group
membership status specifically rather than group by group.

Loading the groups in bulk, and only optionally augmenting them with
owner and manager information, removes the need for many database
queries: before the /assignments endpoint required
n_groups x (2 + n_managers_per_group) queries to load group data, from
which only the group name was used. Now it requires just 1 query since
most of the information was not needed and can be skipped.

We also then filtered the assignments by the user's group membership
status, but the only way to do this was to load the statuses for
_all_ users in a group, for each group. This adds a new endpoint to load
a specific user's group status for all their groups, reducing the number
of queries required down to always 1.

For /assignments this reduces the total number of queries from
~ (n_groups x (3 + n_managers_per_group)) + n_teachers + 4
down to
6 + n_teachers
which is a vast improvement!