Fix GroupManager tests using a partial mock
See comment in IsaacTest for details, but in order to actually test the
GroupManager::filterItemsBasedOnMembershipContext method, yet mock all
the other methods, we need some funky partial mocking.

The partial mock cannot have replay(...) called on it twice, it seems,
so I had to remove it from the second replay call in QuizFacadeTest.