Only load relevant attempts on question page load
For historic reasons (MongoDB!), when loading a question page we would
load every single question attempt a user had ever made.
We did this because we need both the attempts at the question parts on
the question page, but also the attempts at any related questions.

For registered users we can optimise this.
Instead we now load the question attempts for the specific page in full;
i.e. the complete QuestionValidationResponses. We need these for the
bestAttempt attribute on the question page. For related questions, we
can instead load LightweightQuestionValidationResponses since we don't
need to deserialise all of the JSON to say whether they are complete or
not. We can walk the content and only load the questions for the
relevant question page IDs.
Anonymous users have less question attempts and there are no equivalent
methods for loading a subset of attempts (they still live in one JSON
blob just like MongoDB), so we just load everything for them.

The efficiency gains for highly active users are immense; for a user
with thousands of question attempts, the page load time speedup is 10x.
Most of the gain is from no longer deserialising irrelevant attempts,
converting them back to strings for the ETag calculation.