Do not cache DTO objects in the GitContentManager
DTO objects exist to be transferred to the client, and we were caching
them because it is expensive to load them from ElasticSearch and also
expensive (but less so) to map them from DO to DTO form.

However, despite the warnings in the JavaDoc, we keep mutating the DTOs
returned by the GCM without cloning them. This is understandable; they
exist to contain user data for transfer.
Rather than ditch the cache entirely, we remove the cache for DTOs but
continue to cache DOs, since the latter are actually supposed to be
unchanging anyway and we do not mutate them anywhere at present.
This adds ~1-3ms onto a cache hit, but it prevents all sorts of data
leaking from being possible. Since a cache miss is ~10ms, hopefully this
extra time will be okay.