Fix unwanted null return value from getMatchingQs
There's a .get() in here that takes us from the three-layer map to the
two layer map by getting the specific user of interest. But if there
are no attempts for that user, we returned null and not an empty map.

For efficiency, given all these maps will not be modified, use a static
Collections.emptyMap() object.