Use DOs for assignment and gameboard calculations
In these cases we were not using the question DTOs for transfer, but as
reference for what parts a question contains. We can use DOs for this
since they are the authoritative version of questions, and we do not
plan to use the objects for anything which is transferred to the client.
This avoids the need to create new DTO objects every time, and can use
the DO cache.