Stop using token owner for group join permissions
For some reason both groups and group join tokens have ownership
information. We added support for altering the owner of a group, but
this did not alter the owner of the token.
When users tried to join a group with a changed owner, the new owner was
not listed amongst the users to approve, but the old token owner was.
This meant that the group owner was not given permission.

This commit stops using the token ownership information for anything;
tokens really belong to groups, not to users, and the old approach of
tokens belonging to users has not been correct in some time.
As such, the test that group tokens could work without a group is
not useful; they cannot work that way since the introduction of
additional managers.

This commit also purges methods for creating an association based on a
token, since that method did not include group managers; it was
misleading and required group managers to be dealt with separately.