Notify user if account email change already exists
Currently we implicitly handle this exception as a subclass of the
SegueDatabaseException underneath, which just says "Error while
modifying the user" to the client. It also prints a stack trace, neither
of which are helpful.
This could be used for account enumeration, but we already do the same
thing on registration and there are plenty of other account enumeration
surfaces. This one is not side-effect free, and so cannot easily be
scripted.