Remove credentials of deleted users
After we moved to soft-deletion of users, the rows of this table stopped
being removed by the ON DELETE CASCADE. But we have no cause for keeping
old credentials and so should remove them.
For existing deleted users we can remove the credentials with an SQL
script.
This forms important groundwork for archiving of user accounts.