Use Optional in getContentDOById to cache not found
If we attempt to load a non-existent piece of content, say a page
fragment used optionally in the frontend, then we will fail to find it
in ElasticSearch and return that failure. But repeated requests will
continue to go to ElasticSearch rather than be served by the content
cache, even though we know they will all fail.
Rather than directly cache the Content objects, cache Optionals, which
allow Optional.empty() for the "not found" case.
Also move to the more atomic .get(key, Callable) form to fetch the
missing value if it is absent. This better guarantees that the Optional
really does exist.

Having built this, I am not actually sure it is cleaner than just
having a marker singleton for "not found".