Enforce new password security with Pwned Passwords
This uses the HIBP Pwned Passwords k-anonymity API to check how many
times a password has been seen in a data breach and restrict the usage
of common passwords.
It requires a config property for the API URL, and allows an additional
config property to set the denial threshold from the default of 1000.
The default is fairly conservative, but the override can be adjusted
at runtime if necessary.

Not completely sure about the class name and location, however.