Fix content indexer for null getPublished() values
The return value of content.getPublished() was Boolean, but it always
returns a boolean (unboxed) value. However, we would like to allow this
to be null, and if we do there are several places that assume they have
the primitive type incorrectly.

Also, setting "published" to be the empty string on an object that will
later (in GitContentManager::getProblemMap) be cast to Boolean was also
always invalid and worked only because that code path was never used.